feat: add owner company reference in building document

+++ src/modules/building/building.service.ts
@@ -4,6 +4,7 @@ import Building from "./building.model";
 import { ApiError } from "../errors";
 import { IOptions, QueryResult } from '../paginate/paginate';
 import { IBuildingDoc, NewCreatedBuilding, UpdateBuildingBody } from "./building.interfaces";
+import { Company } from "../company";
 
 /**
  * Create a building
@@ -15,7 +16,11 @@ export const createBuilding = async (buildingBody: NewCreatedBuilding): Promise<
         throw new ApiError(httpStatus.BAD_REQUEST, 'Building already exists');
     }
 
-    return Building.create(buildingBody);
+    const company = await Company.findOne({ name: buildingBody.ownerCompanyName })
+    if (!company)
+        throw new ApiError(httpStatus.NOT_FOUND, 'Owner company does not exist')
+
+    return Building.create({ ...buildingBody, ownerCompany: { companyName: company.name, companyId: company._id } });
 };
 
 /**
@@ -55,6 +60,16 @@ export const updateBuildingById = async (
         throw new ApiError(httpStatus.BAD_REQUEST, 'Name already taken');
     }
 
+    if (updateBody.ownerCompanyName) {
+        const targetCompany = await Company.findOne({ name: updateBody.ownerCompanyName })
+        if (!targetCompany) {
+            throw new ApiError(httpStatus.NOT_FOUND, 'Company not found')
+        }
+        delete updateBody.ownerCompanyName
+
+        Object.assign(updateBody, { ownerCompany: { companyName: targetCompany.name, companyId: targetCompany._id } })
+    }
+
     Object.assign(building, updateBody);
     await building.save();
     return building;
@@ -70,6 +85,12 @@ export const deleteBuildingById = async (buildingId: mongoose.Types.ObjectId): P
     if (!building) {
         throw new ApiError(httpStatus.NOT_FOUND, 'Building not found');
     }
+
+    const companies = await Company.find({ 'buildings.buildingId': building._id })
+
+    if (companies)
+        throw new ApiError(httpStatus.CONFLICT, 'Company dependency exists. Delete Not possible')
+
     await building.deleteOne();
     return building;
 };
